refactor: extract EPUB manifest evidence
